Global comments clean up & dead code removal
