+++ src/notation/controller/index.ts
@@ -3,8 +3,3 @@ export * from "./element";
 export * from "./selection";
 export * from "./track-controller";
 export * from "./editor-layout-dimensions";
-
-// TODO(P0-ARCH): Revisit top-level controller module boundaries once Phase 0
-// correctness work is done. In particular, check whether `editor/`,
-// `selection/`, and the TrackController façade still express the clearest
-// ownership model.
